requirements.txt
Flask
black
fdb
pyyaml
simplejson
parameterized
coverage
requests
websockets
asyncio
